# order_db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.28.0

package order_db

type Order struct {
	ID        int32
	Symbol    string
	Price     float64
	Quantity  int32
	OrderType string
}